chore(src): reorganize Icon elements
