+++ src/helix-ui/elements/hx-icon/_shadow.less
